requirements.txt
selenium
pytesseract
Pillow
requests
beautifulsoup4
